prefer using int/uint instead of always using int32_t/uint32_t/etc
I believe this was originally based on an overreaction on my part from
several years ago where I found some code I wrote did not work correctly
on 64-bit systems because I wasn't taking into account different type
sizes.

Use of int32_t, uint32_t, int16_t, uint16_t, etc should only be used
where the byte size of the variable is very important. For general
number variable usage, this is less important. e.g. using int or uint
for screen coordinates is fine. Using int16/uint16 for screen
coordinates also works, but is maybe a bit overkill and more to type
for little-to-no benefit.

+++ src/contexts/contentcache.h
@@ -25,7 +25,7 @@ public:
 	SpriteFont* GetFont() const                            { return m_standardFont; }
 	SpriteFont* GetUIFont() const                          { return m_uiFont; }
 	const stl::string& GetUIFontFilename() const           { return m_uiFontFilename; }
-	uint8_t GetUIFontSize() const                          { return m_uiFontSize; }
+	uint GetUIFontSize() const                             { return m_uiFontSize; }
 
 private:
 	ContentManager *m_contentManager;
@@ -36,7 +36,7 @@ private:
 	SpriteFont *m_standardFont;
 	SpriteFont *m_uiFont;
 	stl::string m_uiFontFilename;
-	uint8_t m_uiFontSize;
+	uint m_uiFontSize;
 };
 
 #endif
