+++ src/effects/dimeffect.cpp
@@ -20,8 +20,8 @@ DimEffect::~DimEffect()
 
 void DimEffect::OnRender(RenderContext *renderContext)
 {
-	uint16_t width = renderContext->GetGraphicsDevice()->GetViewContext()->GetViewportWidth();
-	uint16_t height = renderContext->GetGraphicsDevice()->GetViewContext()->GetViewportHeight();
+	uint width = renderContext->GetGraphicsDevice()->GetViewContext()->GetViewportWidth();
+	uint height = renderContext->GetGraphicsDevice()->GetViewContext()->GetViewportHeight();
 	
 	m_color.a = m_alpha;
 	renderContext->GetSpriteBatch()->Render(
