+++ src/framework/graphics/billboardspritebatch.h
@@ -106,7 +106,7 @@ public:
 	 * @param type the type of billboard to render this texture as
 	 * @param color color to tint the texture with
 	 */
-	void Render(const TextureAtlas *atlas, uint32_t index, float x, float y, float z, float width, float height, BILLBOARDSPRITE_TYPE type = BILLBOARDSPRITE_SPHERICAL, const Color &color = COLOR_WHITE);
+	void Render(const TextureAtlas *atlas, uint index, float x, float y, float z, float width, float height, BILLBOARDSPRITE_TYPE type = BILLBOARDSPRITE_SPHERICAL, const Color &color = COLOR_WHITE);
 
 	/**
 	 * Renders a texture atlas sub-texture / tile as a billboard. The billboard 
@@ -118,7 +118,7 @@ public:
 	 * @param type the type of billboard to render this texture as
 	 * @param color color to tint the texture with
 	 */
-	void Render(const TextureAtlas *atlas, uint32_t index, const Vector3 &position, float width, float height, BILLBOARDSPRITE_TYPE type = BILLBOARDSPRITE_SPHERICAL, const Color &color = COLOR_WHITE);
+	void Render(const TextureAtlas *atlas, uint index, const Vector3 &position, float width, float height, BILLBOARDSPRITE_TYPE type = BILLBOARDSPRITE_SPHERICAL, const Color &color = COLOR_WHITE);
 
 	/**
 	 * Renders text as series of billboards. The text will be centered on
@@ -183,20 +183,20 @@ public:
 	void End();
 
 private:
-	void Initialize(GraphicsDevice *graphicsDevice, uint16_t maxSprites);
+	void Initialize(GraphicsDevice *graphicsDevice, uint maxSprites);
 	void InternalBegin(const RenderState *renderState, const BlendState *blendState, SpriteShader *shader);
-	void AddSprite(BILLBOARDSPRITE_TYPE type, const Texture *texture, const Vector3 &position, float width, float height, uint16_t sourceLeft, uint16_t sourceTop, uint16_t sourceRight, uint16_t sourceBottom, const Color &color);
+	void AddSprite(BILLBOARDSPRITE_TYPE type, const Texture *texture, const Vector3 &position, float width, float height, uint sourceLeft, uint sourceTop, uint sourceRight, uint sourceBottom, const Color &color);
 	void AddSprite(BILLBOARDSPRITE_TYPE type, const Texture *texture, const Vector3 &position, float width, float height, float texCoordLeft, float texCoordTop, float texCoordRight, float texCoordBottom, const Color &color);
-	void AddSprite(BILLBOARDSPRITE_TYPE type, const Matrix4x4 &transform, const Texture *texture, const Vector3 &offset, float width, float height, uint16_t sourceLeft, uint16_t sourceTop, uint16_t sourceRight, uint16_t sourceBottom, const Color &color);
+	void AddSprite(BILLBOARDSPRITE_TYPE type, const Matrix4x4 &transform, const Texture *texture, const Vector3 &offset, float width, float height, uint sourceLeft, uint sourceTop, uint sourceRight, uint sourceBottom, const Color &color);
 	void AddSprite(BILLBOARDSPRITE_TYPE type, const Matrix4x4 &transform, const Texture *texture, const Vector3 &offset, float width, float height, float texCoordLeft, float texCoordTop, float texCoordRight, float texCoordBottom, const Color &color);
-	void SetSpriteInfo(uint32_t spriteIndex, BILLBOARDSPRITE_TYPE type, const Matrix4x4 &transform, const Texture *texture, const Vector3 &offset, float width, float height, float texCoordLeft, float texCoordTop, float texCoordRight, float texCoordBottom, const Color &color);
+	void SetSpriteInfo(uint spriteIndex, BILLBOARDSPRITE_TYPE type, const Matrix4x4 &transform, const Texture *texture, const Vector3 &offset, float width, float height, float texCoordLeft, float texCoordTop, float texCoordRight, float texCoordBottom, const Color &color);
 	void RenderQueue();
-	void RenderQueueRange(const Texture *texture, uint32_t firstSprite, uint32_t lastSprite);
+	void RenderQueueRange(const Texture *texture, uint firstSprite, uint lastSprite);
 
 	Matrix4x4 GetTransformFor(BILLBOARDSPRITE_TYPE type, const Vector3 &position) const;
 	void CheckForNewSpriteSpace();
 
-	uint32_t m_currentSpriteCapacity;
+	uint m_currentSpriteCapacity;
 	GraphicsDevice *m_graphicsDevice;
 	SpriteShader *m_shader;
 	RenderState *m_renderState;
@@ -207,7 +207,7 @@ private:
 	BlendState m_overrideBlendState;
 	VertexBuffer *m_vertices;
 	stl::vector<const Texture*> m_textures;
-	uint32_t m_currentSpritePointer;
+	uint m_currentSpritePointer;
 
 	Vector3 m_cameraPosition;
 	Vector3 m_cameraForward;
